+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Master
 
-* Ensure graphml output is correctly ordered (nodes must come before edges). Orientdb json export cannot be guaranteed to order
+* Prune orphaned edges. Neo4j import will fail if it detects edges that reference absent nodes.
+* Ensure graphml output is correctly ordered (nodes must come before edges). OrientDB json export cannot be guaranteed to order
 nodes before edges, but neo4j's import tool requires this.
 
 ## 1.1.0